requirements.txt
Django>=4.1,<4.2
django-environ>=0.7,<1.0
djangorestframework>=3.14.0
django-cors-headers>=3.11.0,<3.12.0